requirements.txt
# Core web framework
flask==2.0.1

# Computer vision and image processing
opencv-python==4.5.3.56
numpy==1.21.2
cvzone==1.5.6

# AI/ML models and inference
ultralytics==8.0.196
paddleocr==2.6.1.3

# Database connectivity
mysql-connector-python==8.1.0

# Data analysis and visualization (for notebooks)
pandas>=1.3.0
matplotlib>=3.4.0
seaborn>=0.11.0

# Jupyter notebook support
jupyter>=1.0.0
ipython>=7.0.0

# Additional utilities
pillow>=8.0.0
tqdm>=4.60.0


